# openapi/uniflow.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FRINX Uniflow OpenAPI",
    "version": "v0"
  },
  "servers": [
    {
      "url": "/api/uniflow",
      "description": "Frinx APi gateway"
    }
  ],
  "paths": {
    "/hierarchical": {
      "get": {
        "tags": [
          "workflow-search"
        ],
        "summary": "Search for workflows based on payload and other parameters",
        "description": "use sort options as sort=<field>:ASC|DESC e.g. sort=name&sort=workflowId:DESC. If order is not specified, defaults to ASC.",
        "operationId": "search",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workflowId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "*"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultWorkflowSummary"
                }
              }
            }
          }
        }
      }
    },
    "/executions": {
      "get": {
        "tags": [
          "workflow-search"
        ],
        "summary": "Search for workflows and tasks based on payload and other parameters",
        "description": "use sort options as sort=<field>:ASC|DESC e.g. sort=name&sort=workflowId:DESC. If order is not specified, defaults to ASC.",
        "operationId": "executions",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workflowId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "*"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultWorkflowSummary"
                }
              }
            }
          }
        }
      }
    },
    "/id/{workflowId}": {
      "get": {
        "tags": [
          "workflow-search"
        ],
        "summary": "Search workflow by workflowId",
        "operationId": "id",
        "parameters": [
          {
            "name": "workflowId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultWorkflowIdSummary"
                }
              }
            }
          }
        }
      }
    },
    "/workflow/{workflowId}": {
      "delete": {
        "tags": [
          "workflow-resource"
        ],
        "summary": "Terminate workflow execution",
        "operationId": "terminate",
        "parameters": [
          {
            "name": "workflowId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/workflow": {
      "post": {
        "tags": [
          "workflow-resource"
        ],
        "summary": "Start a new workflow with StartWorkflowRequest, which allows task to be executed in a domain",
        "operationId": "startWorkflow",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/external/postgres/{externalPayloadPath}": {
      "get": {
        "tags": [
          "external-postgres-payload-resource"
        ],
        "summary": "Get task or workflow by externalPayloadPath from External PostgreSQL Storage",
        "operationId": "getExternalStorageData",
        "parameters": [
          {
            "name": "externalPayloadPath",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/workflow/bulk/terminate": {
      "delete": {
        "tags": [
          "workflow-bulk-resource"
        ],
        "summary": "Terminate workflows execution (deprecated)",
        "operationId": "terminate_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workflow-bulk-resource"
        ],
        "summary": "Terminate workflows execution",
        "operationId": "terminate_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workflow/bulk/pause": {
      "put": {
        "tags": [
          "workflow-bulk-resource"
        ],
        "summary": "Pause the list of workflows",
        "operationId": "pauseWorkflow_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workflow/bulk/resume": {
      "put": {
        "tags": [
          "workflow-bulk-resource"
        ],
        "summary": "Resume the list of workflows",
        "operationId": "resumeWorkflow_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workflow/bulk/restart": {
      "post": {
        "tags": [
          "workflow-bulk-resource"
        ],
        "summary": "Restart the list of completed workflow",
        "operationId": "restart_1",
        "parameters": [
          {
            "name": "useLatestDefinitions",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workflow/bulk/retry": {
      "post": {
        "tags": [
          "workflow-bulk-resource"
        ],
        "summary": "Retry the last failed task for each workflow from the list",
        "operationId": "retry_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SearchResultWorkflowSummary": {
        "type": "object",
        "properties": {
          "totalHits": {
            "type": "integer",
            "format": "int64"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowSummary"
            }
          }
        }
      },
      "SearchResultWorkflowIdSummary": {
        "type": "object",
        "properties": {
          "results": {
            "type": "object",
            "$ref": "#/components/schemas/Workflow"
          }
        }
      },
...
      "StartWorkflowRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "correlationId": {
            "type": "string"
          },
          "input": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          },
          "taskToDomain": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "workflowDef": {
            "$ref": "#/components/schemas/WorkflowDef"
          },
          "externalInputPayloadStoragePath": {
            "type": "string"
          },
          "priority": {
            "maximum": 99,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}